% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scac_database_downloader.R
\name{download.scac.database}
\alias{download.scac.database}
\title{Download the Most Recent Version of the SCAC Filings Database}
\usage{
download.scac.database(start.page = 1, debug = FALSE,
  update.cache = FALSE, cache.file = file.path(system.file(package =
  "scac.database.downloader", "data"), "scacdb.rdata"))
}
\arguments{
\item{start.page}{numeric Which page of the filings html database to start scraping from}

\item{debug}{logical Whether or not to be verbose}

\item{update.cache}{logical If TRUE we download the entire database and update locally cached copy
if FALSE we simply return the locally cached copy}

\item{cache.file}{character path to .rdata file where most recent copy of the scac db
is stored}
}
\value{
returns data.frame
}
\description{
Download the database of securities class action fillings from SCAC
as published at: http://securities.stanford.edu/filings.html
}
