\name{scac.database.downloader-package}
\alias{scac.database.downloader-package}
\alias{scac.database.downloader}
\docType{package}
\title{
\packageTitle{scac.database.downloader}
}
\description{
\packageDescription{scac.database.downloader}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scac.database.downloader}
\packageIndices{scac.database.downloader}

}
\author{
\packageAuthor{scac.database.downloader}

Maintainer: \packageMaintainer{scac.database.downloader}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
