% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scac.db}
\alias{scac.db}
\title{The Securities Class Action Clearinghouse (SCAC)}
\format{A data.frame with 3000+ rows and 14 variables
\itemize{
\item FilingName: name of the filing
\item FilingDate: filing date
\item DistrictCourt: court 
\item Exchange: exchange
\item Ticker: ticker symbol
\item CaseSummary: summmary
\item Sector: sector
\item Industry: industry
\item Headquarters: headquarters location
\item FirstIdentifiedComplainant: first identified complainant
\item Judge: presiding judge
\item PlaintiffFirms: plaintiff law firms
\item CaseLink: link to full case details
\item FetchDate: date record was last fetched on
}}
\source{
The Securities Class Action Clearinghouse (SCAC) \url{http://securities.stanford.edu/}
}
\usage{
scac.db
}
\description{
The Securities Class Action Clearinghouse (SCAC) provides detailed information relating to the prosecution, defense, and settlement of federal class action securities fraud litigation.

The SCAC team maintains a Filings database of more than 4,000 securities class action lawsuits filed since passage of the Private Securities Litigation Reform Act of 1995. The database also contains copies of complaints, briefs, dispositive orders, and other litigation-related materials filed in these cases.

The SCAC offers regular email notifications of new securities class action filings and news. Institutional investors with more than $1.5 trillion in assets under management are registered with the SCAC to receive automatic notices of litigation developments that may affect investments in their portfolios. Hundreds of attorneys, executives, scholars, and media personnel have also registered with the SCAC. If you would like to receive email notifications, please register \url{http://securities.stanford.edu/about-the-scac.html#register}.
}
\keyword{datasets}